[metricbeat/module/prometheus] [metricbeat/module/openmetrics] Keep histogram CumulativeCount and SampleCount as float64 internally to avoid undefined behavior when checking for NaN and Inf

+++ metricbeat/module/openmetrics/collector/collector_test.go
@@ -122,7 +122,7 @@ func TestGetOpenMetricsEventsFromMetricFamily(t *testing.T) {
 				Metric: []*p.OpenMetric{
 					{
 						Summary: &p.Summary{
-							SampleCount: proto.Uint64(10),
+							SampleCount: proto.Float64(10),
 							SampleSum:   proto.Float64(10),
 							Quantile: []*p.Quantile{
 								{
@@ -166,12 +166,12 @@ func TestGetOpenMetricsEventsFromMetricFamily(t *testing.T) {
 				Metric: []*p.OpenMetric{
 					{
 						Histogram: &p.Histogram{
-							SampleCount: proto.Uint64(10),
+							SampleCount: proto.Float64(10),
 							SampleSum:   proto.Float64(10),
 							Bucket: []*p.Bucket{
 								{
 									UpperBound:      proto.Float64(0.99),
-									CumulativeCount: proto.Uint64(10),
+									CumulativeCount: proto.Float64(10),
 								},
 							},
 						},
@@ -299,7 +299,7 @@ func TestSkipMetricFamily(t *testing.T) {
 			Metric: []*p.OpenMetric{
 				{
 					Summary: &p.Summary{
-						SampleCount: proto.Uint64(10),
+						SampleCount: proto.Float64(10),
 						SampleSum:   proto.Float64(10),
 						Quantile: []*p.Quantile{
 							{
@@ -318,12 +318,12 @@ func TestSkipMetricFamily(t *testing.T) {
 			Metric: []*p.OpenMetric{
 				{
 					Histogram: &p.Histogram{
-						SampleCount: proto.Uint64(10),
+						SampleCount: proto.Float64(10),
 						SampleSum:   proto.Float64(10),
 						Bucket: []*p.Bucket{
 							{
 								UpperBound:      proto.Float64(0.99),
-								CumulativeCount: proto.Uint64(10),
+								CumulativeCount: proto.Float64(10),
 							},
 						},
 					},
