+++ metricbeat/module/prometheus/collector/collector_test.go
@@ -110,7 +110,7 @@ func TestGetPromEventsFromMetricFamily(t *testing.T) {
 				Metric: []*p.OpenMetric{
 					{
 						Summary: &p.Summary{
-							SampleCount: proto.Uint64(10),
+							SampleCount: proto.Float64(10),
 							SampleSum:   proto.Float64(10),
 							Quantile: []*p.Quantile{
 								{
@@ -152,12 +152,12 @@ func TestGetPromEventsFromMetricFamily(t *testing.T) {
 				Metric: []*p.OpenMetric{
 					{
 						Histogram: &p.Histogram{
-							SampleCount: proto.Uint64(10),
+							SampleCount: proto.Float64(10),
 							SampleSum:   proto.Float64(10),
 							Bucket: []*p.Bucket{
 								{
 									UpperBound:      proto.Float64(0.99),
-									CumulativeCount: proto.Uint64(10),
+									CumulativeCount: proto.Float64(10),
 								},
 							},
 						},
@@ -280,7 +280,7 @@ func TestSkipMetricFamily(t *testing.T) {
 			Metric: []*p.OpenMetric{
 				{
 					Summary: &p.Summary{
-						SampleCount: proto.Uint64(10),
+						SampleCount: proto.Float64(10),
 						SampleSum:   proto.Float64(10),
 						Quantile: []*p.Quantile{
 							{
@@ -299,12 +299,12 @@ func TestSkipMetricFamily(t *testing.T) {
 			Metric: []*p.OpenMetric{
 				{
 					Histogram: &p.Histogram{
-						SampleCount: proto.Uint64(10),
+						SampleCount: proto.Float64(10),
 						SampleSum:   proto.Float64(10),
 						Bucket: []*p.Bucket{
 							{
 								UpperBound:      proto.Float64(0.99),
-								CumulativeCount: proto.Uint64(10),
+								CumulativeCount: proto.Float64(10),
 							},
 						},
 					},
